temp commit

+++ include/nanofs/block.h
@@ -7,15 +7,15 @@
 /* Block level interface                                                    */
 /* ======================================================================== */
 
-void fs_low_level_storage fs_block_free(uint32_t dev_block_addr);
+void fs_low_level_storage fs_block_free(uint16_t dev_block_addr);
 
-uint32_t fs_low_level_storage fs_block_alloc();
+uint8_t fs_low_level_storage fs_block_alloc();
 
 void fs_low_level_storage block_init(void);
 
-void block_read(char *dst, int block_num);
+void block_read(char *dst, uint8_t block_num);
 
-void block_write(char *src, int block_num);
+void block_write(char *src, uint8_t block_num);
 
 
 #endif // NANOFS_BLOCK_H
